+++ R/tree_distance.R
@@ -93,63 +93,15 @@
 #' @importFrom clue solve_LSAP
 #' @export
 MutualArborealInfo <- function (tree1, tree2, reportMatching = FALSE) {
-  if (class(tree1) == 'phylo') {
-    if (class(tree2) == 'phylo') {
-      if (length(setdiff(tree1$tip.label, tree2$tip.label)) > 0) {
-        stop("Tree tips must bear identical labels")
-      }
-      
-      MutualArborealInfoSplits(Tree2Splits(tree1), Tree2Splits(tree2), reportMatching)
-    } else {
-      splits1 <- Tree2Splits(tree1)
-      vapply(tree2, 
-             function (tr2) MutualArborealInfoSplits(splits1, Tree2Splits(tr2)),
-             double(1))
-    }
-  } else {
-    if (class(tree2) == 'phylo') {
-      splits1 <- Tree2Splits(tree2)
-      vapply(tree1, 
-             function (tr2) MutualArborealInfoSplits(splits1, Tree2Splits(tr2)),
-             double(1))
-    } else {
-      splits1 <- lapply(tree1, Tree2Splits)
-      splits2 <- lapply(tree2, Tree2Splits)
-      matrix(mapply(MutualArborealInfoSplits, rep(splits1, each=length(splits2)), splits2),
-             length(splits2), length(splits1), dimnames = list(names(tree2), names(tree1)))
-    }
-  }
+  CalculateTreeDistance(MutualArborealInfoSplits, tree1, tree2, 
+                        reportMatching=reportMatching)
 }
 
 #' @describeIn MutualArborealInfo Variation of phylogenetic information between two trees
 #' @export
 VariationOfArborealInfo <- function (tree1, tree2, reportMatching = FALSE) {
-  if (class(tree1) == 'phylo') {
-    if (class(tree2) == 'phylo') {
-      if (length(setdiff(tree1$tip.label, tree2$tip.label)) > 0) {
-        stop("Tree tips must bear identical labels")
-      }
-      
-      VariationOfArborealInfoSplits(Tree2Splits(tree1), Tree2Splits(tree2), reportMatching)
-    } else {
-      splits1 <- Tree2Splits(tree1)
-      vapply(tree2, 
-             function (tr2) VariationOfArborealInfoSplits(splits1, Tree2Splits(tr2)),
-             double(1))
-    }
-  } else {
-    if (class(tree2) == 'phylo') {
-      splits1 <- Tree2Splits(tree2)
-      vapply(tree1, 
-             function (tr2) VariationOfArborealInfoSplits(splits1, Tree2Splits(tr2)),
-             double(1))
-    } else {
-      splits1 <- lapply(tree1, Tree2Splits)
-      splits2 <- lapply(tree2, Tree2Splits)
-      matrix(mapply(VariationOfArborealInfoSplits, rep(splits1, each=length(splits2)), splits2),
-             length(splits2), length(splits1), dimnames = list(names(tree2), names(tree1)))
-    }
-  }
+  CalculateTreeDistance(VariationOfArborealInfoSplits, tree1, tree2, 
+                        reportMatching=reportMatching)
 }
 
 #' @describeIn MutualArborealInfo Mutual clustering information between two trees
@@ -157,41 +109,84 @@ VariationOfArborealInfo <- function (tree1, tree2, reportMatching = FALSE) {
 MutualClusterInfo <- function (tree1, tree2,
                                   reportMatching = FALSE,
                                   bestMatchOnly = TRUE) {
+  CalculateTreeDistance(MutualClusterInfoSplits, tree1, tree2, 
+                        reportMatching=reportMatching, bestMatchOnly=bestMatchOnly)
+}
+
+#' Nye et al. (2006) tree comparison
+#' 
+#' Implements the tree comparison metric of Nye _et al_. (2006).
+#' 
+#' @inheritParams MutualArborealInfo
+#' 
+#' @references \insertRef{Nye2006}{TreeSearch}
+#' @concept Tree distance
+#' 
+#' @author Martin R. Smith
+#' @export
+NyeTreeSimilarity <- function (tree1, tree2,
+                             reportMatching = FALSE) {
+  CalculateTreeDistance(NyeSplitSimilarity, tree1, tree2, reportMatching)
+}
+
+#' Matching Split Distance
+#' 
+#' Implements the Matching Split Distance for unrooted binary phylogenetic 
+#' trees of Bogdanowicz and Giaro (2012).
+#' 
+#' @inheritParams MutualArborealInfo
+#' 
+#' @references \insertRef{Bogdanowicz2012}{TreeSearch}
+#' @concept Tree distance
+#' 
+#' @author Martin R. Smith
+#' @export
+MatchingSplitDistance <- function (tree1, tree2,
+                             reportMatching = FALSE) {
+  CalculateTreeDistance(MatchingSplitDistanceSplits, tree1, tree2, reportMatching)
+}
+
+#' Wrapper for tree distance calculations
+#' 
+#' Calls tree distance functions from trees or lists of trees
+#' 
+#' @inheritParams MutualArborealInfo
+#' @param Func Tree distance function.
+#' @param \dots Additional arguments to `Func`.
+#' 
+#' @author Martin R. Smith
+#' @keywords internal
+#' @export
+CalculateTreeDistance <- function (Func, tree1, tree2, reportMatching, ...) {
   if (class(tree1) == 'phylo') {
     if (class(tree2) == 'phylo') {
       if (length(setdiff(tree1$tip.label, tree2$tip.label)) > 0) {
         stop("Tree tips must bear identical labels")
       }
       
-      MutualClusterInfoSplits(Tree2Splits(tree1), Tree2Splits(tree2),
-                                    reportMatching, bestMatchOnly)
+      Func(Tree2Splits(tree1), Tree2Splits(tree2), reportMatching, ...)
     } else {
       splits1 <- Tree2Splits(tree1)
       vapply(tree2, 
-             function (tr2) MutualClusterInfoSplits(splits1, Tree2Splits(tr2),
-                                                      reportMatching,
-                                                      bestMatchOnly),
+             function (tr2) Func(splits1, Tree2Splits(tr2), ...),
              double(1))
     }
   } else {
     if (class(tree2) == 'phylo') {
       splits1 <- Tree2Splits(tree2)
       vapply(tree1, 
-             function (tr2) MutualClusterInfoSplits(splits1, Tree2Splits(tr2),
-                                                          reportMatching, bestMatchOnly),
+             function (tr2) Func(splits1, Tree2Splits(tr2), ...),
              double(1))
     } else {
       splits1 <- lapply(tree1, Tree2Splits)
       splits2 <- lapply(tree2, Tree2Splits)
-      matrix(mapply(MutualClusterInfoSplits,
-                    rep(splits1, each=length(splits2)),
-                    splits2,
-                    reportMatching=reportMatching,
-                    bestMatchOnly= bestMatchOnly),
-             length(splits2), length(splits1), dimnames = list(names(tree2), names(tree1)))
+      matrix(mapply(Func, rep(splits1, each=length(splits2)), splits2),
+             length(splits2), length(splits1),
+             dimnames = list(names(tree2), names(tree1)), ...)
     }
   }
 }
+
 
 #' @describeIn MutualArborealInfo Takes splits instead of trees
 #' @export
@@ -400,6 +395,142 @@ VariationOfArborealInfoSplits <- function (splits1, splits2, reportMatching = FA
   }
 }
 
+#' @describeIn NyeTreeSimilarity Takes splits instead of trees
+#' @inheritParams MutualArborealInfoSplits
+#' @export
+NyeSplitSimilarity <- function (splits1, splits2, reportMatching = FALSE) {
+  
+  dimSplits1 <- dim(splits1)
+  dimSplits2 <- dim(splits2)
+  nTerminals <- dimSplits1[1]
+  if (dimSplits2[1] != nTerminals) {
+    stop("Split rows must bear identical labels")
+  }
+  
+  if (dimSplits1[2] < dimSplits2[2]) {
+    tmp <- splits1
+    splits1 <- splits2
+    splits2 <- tmp
+    
+    tmp <- dimSplits1
+    dimSplits1 <- dimSplits2
+    dimSplits2 <- tmp
+  }
+  
+  taxonNames <- rownames(splits1) 
+  
+  if (!is.null(taxonNames)) {
+    splits2 <- unname(splits2[rownames(splits1), , drop=FALSE])
+    splits1 <- unname(splits1) # split1[split2] faster without names
+  }
+  
+  nSplits1 <- dimSplits1[2]
+  nSplits2 <- dimSplits2[2]
+  
+  Ars <- function (pir, pjs) {
+    sum(pir[pjs]) / sum(pir | pjs)
+  }
+  
+  pairScores <- matrix((mapply(function(i, j) {
+    splitI0 <- splits1[, i]
+    splitJ0 <- splits2[, j]
+    splitI1 <- !splitI0
+    splitJ1 <- !splitJ0
+    
+    max(
+      min(Ars(splitI0, splitJ0), Ars(splitI1, splitJ1)),
+      min(Ars(splitI0, splitJ1), Ars(splitI1, splitJ0))
+    )
+      
+  }, rep(seq_len(nSplits1), each=nSplits2), seq_len(nSplits2)
+  )), nSplits2, nSplits1)
+  
+  if (nSplits2 == 1) {
+    min(pairScores)
+  } else {
+    optimalMatching <- solve_LSAP(pairScores, TRUE)
+    
+    # Return:
+    ret <- sum(pairScores[matrix(c(seq_along(optimalMatching), optimalMatching), ncol=2L)])
+    if (reportMatching) {
+      attr(ret, 'matching') <- optimalMatching
+      attr(ret, 'pairScores') <- pairScores
+      ret
+    } else {
+      ret
+    }
+  }
+}
+
+#' @describeIn MatchingSplitDistance Takes splits instead of trees
+#' @inheritParams MutualArborealInfoSplits
+#' @export
+MatchingSplitDistanceSplits <- function (splits1, splits2, reportMatching = FALSE) {
+  
+  dimSplits1 <- dim(splits1)
+  dimSplits2 <- dim(splits2)
+  nTerminals <- dimSplits1[1]
+  if (dimSplits2[1] != nTerminals) {
+    stop("Split rows must bear identical labels")
+  }
+  
+  if (dimSplits1[2] < dimSplits2[2]) {
+    tmp <- splits1
+    splits1 <- splits2
+    splits2 <- tmp
+    
+    tmp <- dimSplits1
+    dimSplits1 <- dimSplits2
+    dimSplits2 <- tmp
+  }
+  
+  taxonNames <- rownames(splits1) 
+  
+  if (!is.null(taxonNames)) {
+    splits2 <- unname(splits2[rownames(splits1), , drop=FALSE])
+    splits1 <- unname(splits1) # split1[split2] faster without names
+  }
+  
+  nSplits1 <- dimSplits1[2]
+  nSplits2 <- dimSplits2[2]
+  
+  SymmetricDifference <- function (A, B) {
+    (A & !B) | (!A & B)
+    #TODO: Assert: These will always be the same size
+    #Therefore: Improve efficiency by only calculating one, and not dividing by 2
+  }
+  
+  pairScores <- matrix((mapply(function(i, j) {
+    A1 <- splits1[, i]
+    A2 <- splits2[, j]
+    B1 <- !A1
+    B2 <- !A2
+    
+    min(
+      sum(SymmetricDifference(A1, A2), SymmetricDifference(B1, B2)),
+      sum(SymmetricDifference(A1, A2), SymmetricDifference(B1, B2))
+    ) / 2L
+      
+  }, rep(seq_len(nSplits1), each=nSplits2), seq_len(nSplits2)
+  )), nSplits2, nSplits1)
+  
+  if (nSplits2 == 1) {
+    min(pairScores)
+  } else {
+    optimalMatching <- solve_LSAP(pairScores, FALSE)
+    
+    # Return:
+    ret <- sum(pairScores[matrix(c(seq_along(optimalMatching), optimalMatching), ncol=2L)])
+    if (reportMatching) {
+      attr(ret, 'matching') <- optimalMatching
+      attr(ret, 'pairScores') <- pairScores
+      ret
+    } else {
+      ret
+    }
+  }
+}
+
 #' @describeIn MutualArborealInfo Takes splits instead of trees
 #' 
 #' @param partitionQualityIndex Output of [`SplitPairingInformationIndex`] for
@@ -502,4 +633,60 @@ SplitsCompatible <- function (split1, split2) {
     all(!split1[split2]) ||
     all(!split1[!split2])
   )
+}
+
+#' Visualise a matching
+#' 
+#' Depicts the bipartitions that are matched between two trees using a 
+#' specified Generalized Robinson Foulds tree distance measure.
+#' 
+#' @param Func Function used to construct tree similarity.
+#' @param tree1,tree2 Trees of class `phylo`, with tips labelled identically.
+#' @param setPar Logical specifying whether graphical parameters should be 
+#' set to display trees side by side.
+#' @param Plot Function to use to plot trees.
+#' @param \dots Additional parameters to send to `Plot`.
+#' 
+#' @author Martin R. Smith
+#' @importFrom ape nodelabels edgelabels plot.phylo
+#' @importFrom colorspace qualitative_hcl
+#' @importFrom graphics par
+#' @export
+VisualizeMatching <- function(Func, tree1, tree2, setPar = TRUE, Plot = plot.phylo, ...) {
+  
+  splits1 <- Tree2Splits(tree1)
+  edge1 <- tree1$edge
+  child1 <- edge1[, 2]
+  partitionEdges1 <- vapply(colnames(splits1), 
+                            function (node) which(child1 == node), integer(1))
+  
+  splits2 <- Tree2Splits(tree2)
+  edge2 <- tree2$edge
+  child2 <- edge2[, 2]
+  partitionEdges2 <- vapply(colnames(splits2), 
+                            function (node) which(child2 == node), integer(1))
+  
+  matching <- Func(tree1, tree2, reportMatching = TRUE)
+  pairings <- attr(matching, 'matching')
+  scores <- attr(matching, 'pairScores')
+  pairScores <- signif(mapply(function (i, j) scores[i, j], seq_along(pairings), pairings), 3)
+  
+  nSplits <- length(pairings)
+  palette <- qualitative_hcl(nSplits, c=42, l=88)
+  adjNo <- c(0.5, -0.2)
+  adjVal <- c(0.5, 1.1)
+  
+  if (setPar) origPar <- par(mfrow=c(2, 1), mar=rep(0.5, 4))
+  
+  Plot(tree1)#, ...)
+  edgelabels(seq_along(pairings), partitionEdges1[pairings], bg=palette, adj=adjNo)
+  #edgelabels(seq_along(pairings), partitionEdges1, cex=0.8, font=2, frame='n', adj=adjVal)
+  edgelabels(pairScores, partitionEdges1[pairings], frame='n', adj=adjVal, cex=0.8)
+  
+  Plot(tree2)#, ...)
+  edgelabels(seq_along(pairings), partitionEdges2, bg=palette, adj=adjNo)
+  #edgelabels(seq_along(pairings), partitionEdges2, cex=0.8, font=2, frame='n', adj=adjVal)
+  edgelabels(pairScores, partitionEdges2, frame='n', adj=adjVal, cex=0.8)
+  
+  if (setPar) par(origPar)
 }